+++ jenkins_pipelines/scripts/json_generator/smash_client.py
@@ -61,16 +61,4 @@ class SmashClient():
         json_content : list[dict] = res.json()
         return json_content['results']
     
-    def _issue_to_bsc_link(self, issue: dict[str, any]) -> str:
-        for ref in issue['references']:
-            if ref['source'] == 'SUSE Bugzilla':
-                bsc_num: str = ref['name'].split("#")[1]
-                return f"- [ ] [Bug {bsc_num}]({ref['url']}) - {issue['summary']}\n"
-        raise ValueError(f"No Bugzilla reference for issue {issue['name']}- {issue['summary']}")
-        
-
-    def get_bsc_links_list(self, **kwargs) -> list[str]:
-        issues: list[dict] = self.get_issues(**kwargs)
-
-        return [ self._issue_to_bsc_link(issue) for issue in issues ]
